Extract match stats helper methods
